+++ backend/app/src/main/java/ch/zhaw/studyflow/webserver/controllers/routing/RouteSegment.java
@@ -1,10 +1,5 @@
 package ch.zhaw.studyflow.webserver.controllers.routing;
 
-import ch.zhaw.studyflow.webserver.Tuple;
-
-import java.util.ArrayList;
-import java.util.List;
-
 public record RouteSegment(SegmentType type, String value) {
     public boolean is(SegmentType type) {
         return this.type == type;
