+++ backend/app/src/main/java/ch/zhaw/studyflow/webserver/controllers/routing/RouteTrie.java
@@ -1,7 +1,7 @@
 package ch.zhaw.studyflow.webserver.controllers.routing;
 
 import ch.zhaw.studyflow.webserver.http.HttpMethod;
-import ch.zhaw.studyflow.webserver.Tuple;
+import ch.zhaw.studyflow.utils.Tuple;
 import ch.zhaw.studyflow.webserver.controllers.EndpointMetadata;
 
 import java.util.*;
