+++ backend/app/src/test/java/ch/zhaw/studyflow/webserver/routing/RouteTrieTest.java
@@ -1,7 +1,7 @@
 package ch.zhaw.studyflow.webserver.routing;
 
 import ch.zhaw.studyflow.webserver.http.HttpMethod;
-import ch.zhaw.studyflow.webserver.Tuple;
+import ch.zhaw.studyflow.utils.Tuple;
 import ch.zhaw.studyflow.webserver.controllers.EndpointMetadata;
 import ch.zhaw.studyflow.webserver.controllers.routing.RestRoute;
 import ch.zhaw.studyflow.webserver.controllers.routing.RouteTrie;
